estudos/api/requirements.txt
djangorestframework
markdown
django-filter
